requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filetype==1.0.7
mysqlclient==2.0.3
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
